+++ juriscraper/pacer/__init__.py
@@ -19,6 +19,7 @@ from .list_of_creditors import ListOfCreditors
 from .mobile_query import MobileQuery
 from .rss_feeds import PacerRssFeed
 from .scotus_docket_report import SCOTUSDocketReport
+from .scotus_docket_report_html import SCOTUSDocketReportHTML
 
 __all__ = [
     AcmsCaseSearch,
@@ -45,4 +46,5 @@ __all__ = [
     PossibleCaseNumberApi,
     ShowCaseDocApi,
     SCOTUSDocketReport,
+    SCOTUSDocketReportHTML,
 ]