+++ tests/local/test_ScotusParseDocketTest.py
@@ -1,6 +1,6 @@
 import os
 
-from juriscraper.pacer import SCOTUSDocketReport
+from juriscraper.pacer import SCOTUSDocketReport, SCOTUSDocketReportHTML
 from tests import TESTS_ROOT_EXAMPLES_PACER
 from tests.local.PacerParseTestCase import PacerParseTestCase
 
@@ -16,3 +16,9 @@ class ScotusParseDocketTest(PacerParseTestCase):
             TESTS_ROOT_EXAMPLES_PACER, "dockets", "scotus", "json"
         )
         self.parse_files(path_root, "*.scotus_json", SCOTUSDocketReport)
+
+    def test_parsing_scotus_html_dockets(self):
+        path_root = os.path.join(
+            TESTS_ROOT_EXAMPLES_PACER, "dockets", "scotus", "html"
+        )
+        self.parse_files(path_root, "*.html", SCOTUSDocketReportHTML)